refactor(web): migrate shared context menus and bulk actions to folder convention
