+++ apps/web/src/components/shared/AddToPlaylistButton/index.ts
@@ -1,0 +1,2 @@
+export { default as AddToPlaylistButton } from './AddToPlaylistButton';
+export * from './AddToPlaylistButton.types';
