+++ apps/web/src/components/shared/BulkActionBar/index.ts
@@ -1,0 +1,2 @@
+export { default as BulkActionBar } from './BulkActionBar';
+export * from './BulkActionBar.types';
