+++ apps/web/src/components/shared/PlaylistContextMenu/index.ts
@@ -1,0 +1,2 @@
+export { default as PlaylistContextMenu } from './PlaylistContextMenu';
+export * from './PlaylistContextMenu.types';
