+++ apps/web/src/components/shared/PlaylistPickerContent/index.ts
@@ -1,0 +1,2 @@
+export { default as PlaylistPickerContent } from './PlaylistPickerContent';
+export * from './PlaylistPickerContent.types';
