+++ apps/web/src/components/shared/TrackContextMenu/index.ts
@@ -1,0 +1,2 @@
+export { default as TrackContextMenu } from './TrackContextMenu';
+export * from './TrackContextMenu.types';
